USB: gadeget: ci13xxx: Before remote wake_up cancel suspend work
During suspend of the phy there is a window where clocks are off and
we are still not in LPM. if a remote wakeup happens in this window,
it might lead to unclocked access from the driver. Hence flush the
otg driver suspend work before proceeding with remote wakeup handling.

Signed-off-by: Tarun Gupta <tarung@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>